+++ pom.xml
@@ -114,6 +114,13 @@
 			<scope>test</scope>
 		</dependency>
 
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>5.3.1</version>
+			<scope>test</scope>
+		</dependency>
+
 		<!-- war 배포 시 해당 주석 제거
 		<dependency>
 			<groupId>org.apache.tomcat.embed</groupId>
